requirements.txt
# Core dependencies
discord.py>=2.0.0
python-dotenv>=0.19.0
algoliasearch>=4.0.0
requests>=2.28.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
flask